app/ai-service-mvp/backend/requirements.txt
# AI Service Marketplace - Backend Requirements
# Python 3.11+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# AI/ML
openai==1.3.7
anthropic==0.7.7
langchain==0.0.340
langchain-openai==0.0.2

# Computer Vision
pillow==10.1.0
opencv-python==4.8.1.78

# Telegram
python-telegram-bot==20.7

# Payment Processing
yookassa==2.3.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
phonenumbers==8.13.27

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Task Queue
celery[redis]==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Location Services
geopy==2.4.1

# Date/Time
python-dateutil==2.8.2
pytz==2023.3
